Update 1.19.4 and below 'convert_recipes.py'
We need a trigger to have the advancements work properly (I tried the old version on 1.19.1 and 1.16.5, it does'nt work at all...).

We keep the idea to reveal a recipe only if the player crafts said recipe. But below 1.20-ish, the 'recipe_crafted' trigger does not exist.
We have a way around with the existing 'recipe_unlocked' trigger, which is usually used for 'doLimitedCrafting' servers. This trigger normally triggers if a player is given a special recipe book. But it also triggers if the player crafts for the first time a recipe (thus, this will datapack will never work for 'doLimitedCrafting' server, but those servers might not need this datapack).